Modularizing database models.
