+++ database/DocumentModel.js
@@ -1,0 +1,16 @@
+var mongoose = require('mongoose');
+var Schema = mongoose.Schema;
+
+var documentSchema = new Schema ({
+    id: Number, 
+    title: String, 
+    mode: String,
+    displayName: String, 
+    code: String
+});
+
+var Document = mongoose.model('Document', documentSchema);
+
+module.exports = {
+  userDocument: Document,
+};
