+++ database/MessageModel.js
@@ -1,13 +1,18 @@
-// var mongoose = require('mongoose');
-// var Schema = mongoose.Schema;
+var mongoose = require('mongoose');
+var Schema = mongoose.Schema;
 
-// var messageSchema = new Schema({
-//     nameOfChat: String, 
-//     messageContent: String
-// });
+var messageSchema = new Schema({
+    created: String,
+    text: String,
+    displayName: String,
+    room: String,
+    userid: { type: Number, ref: 'User' }
+});
+
+var Message = mongoose.model('Messages', messageSchema);
 
 
-// var Message = mongoose.model('Messages', messageSchema);
 
-
-// module.exports = Message;
+module.exports = {
+    messages: Message
+};
