+++ database/UserModel.js
@@ -4,8 +4,8 @@ var Schema = mongoose.Schema;
 
 
 
-var uri = config.MONGO_URI; 
-mongoose.connect(uri);
+// var uri = config.MONGO_URI; 
+// mongoose.connect(uri);
 
 var db = mongoose.connection;
 db.on('error', function(err){
@@ -30,42 +30,42 @@ var userSchema = new Schema({
  //this method would probably be best suited for our messages where skills would be messages
  //and the 
 
-var skillsSchema = new mongoose.Schema({
-  node: Number,
-  angular: Number,
-  html: Number,
-  css: Number, 
-  jquery: Number,
-  userid: { type: Number, ref: 'User' }
-});
+// var skillsSchema = new mongoose.Schema({
+//   node: Number,
+//   angular: Number,
+//   html: Number,
+//   css: Number, 
+//   jquery: Number,
+//   userid: { type: Number, ref: 'User' }
+// });
 
-var messageSchema = new Schema({
-    created: String,
-    text: String,
-    displayName: String,
-    room: String,
-    userid: { type: Number, ref: 'User' }
-});
+// var messageSchema = new Schema({
+//     created: String,
+//     text: String,
+//     displayName: String,
+//     room: String,
+//     userid: { type: Number, ref: 'User' }
+// });
 
-var documentSchema = new Schema ({
-    id: Number, 
-    title: String, 
-    mode: String,
-    displayName: String, 
-    code: String
-});
+// var documentSchema = new Schema ({
+//     id: Number, 
+//     title: String, 
+//     mode: String,
+//     displayName: String, 
+//     code: String
+// });
 
-var Document = mongoose.model('Document', documentSchema);
-var Message = mongoose.model('Messages', messageSchema);
-var Skills = mongoose.model("Skills", skillsSchema);
+// var Document = mongoose.model('Document', documentSchema);
+// var Message = mongoose.model('Messages', messageSchema);
+// var Skills = mongoose.model("Skills", skillsSchema);
 var User = mongoose.model("User", userSchema);
 
 
 module.exports = {
-    user: User,
-    messages: Message,
-    userDocument: Document,
-    skills: Skills
+  user: User
+  // messages: Message
+    // userDocument: Document,
+    // skills: Skills
 };
